HttpSession logic added for all methods!

+++ src/main/java/com/banditdev/eksamensprojekt/controller/ProjectController.java
@@ -36,9 +36,9 @@ public class ProjectController {
 
     @GetMapping("myProjects")
     public String showOwnProjects(HttpSession session, Model model) {
-        User currentLoggedInUser = (User) session.getAttribute("user");
 
-        if (currentLoggedInUser == null) {
+        User currentLoggedInUser = (User) session.getAttribute("user");
+        if (!userService.isUserLoggedIn(currentLoggedInUser)) {
             return "redirect:/profile/login";
         }
 
@@ -53,9 +53,11 @@ public class ProjectController {
     public String showProject(@PathVariable int projectId, HttpSession session, Model model) {
 
         User currentLoggedInUser = (User) session.getAttribute("user");
+        if (!userService.isUserLoggedIn(currentLoggedInUser)) {
+            return "redirect:/profile/login";
+        }
 
         Project project = projectService.findProjectById(projectId);
-
         if (project == null) {
             return "redirect:/projects/myProjects";
         }
@@ -74,7 +76,12 @@ public class ProjectController {
 
 
     @GetMapping("/add")
-    public String showAddProjectForm(Model model) {
+    public String showAddProjectForm(Model model, HttpSession session) {
+
+        User currentLoggedInUser = (User) session.getAttribute("user");
+        if (!userService.isUserLoggedIn(currentLoggedInUser)) {
+            return "redirect:/profile/login";
+        }
 
         model.addAttribute("project", new Project());
         model.addAttribute("allUsers",userService.findAllUsers());
@@ -85,9 +92,12 @@ public class ProjectController {
     @PostMapping("/add")
     public String addProject(@ModelAttribute Project project, @RequestParam(required = false) List<Integer> listOfUserIdsFromAssignedUsers, HttpSession session) {
 
-        User currentUser = (User) session.getAttribute("user");
+        User currentLoggedInUser = (User) session.getAttribute("user");
+        if (!userService.isUserLoggedIn(currentLoggedInUser)) {
+            return "redirect:/profile/login";
+        }
 
-        Project createdProject = projectService.addProject(project, currentUser.getUserId());
+        Project createdProject = projectService.addProject(project, currentLoggedInUser.getUserId());
         projectService.addAssignedUserIdsToDatabase(createdProject.getProjectId(), listOfUserIdsFromAssignedUsers);
 
         return "redirect:/projects/" + createdProject.getProjectId();
@@ -96,10 +106,9 @@ public class ProjectController {
     @PostMapping("/delete/{projectId}")
     public String deleteProject(@PathVariable int projectId, HttpSession session) {
 
-        User user = (User) session.getAttribute("user");
-
-        if (user == null) {
-            return "redirect:user/login";
+        User currentLoggedInUser = (User) session.getAttribute("user");
+        if (!userService.isUserLoggedIn(currentLoggedInUser)) {
+            return "redirect:/profile/login";
         }
 
         projectService.removeAllUsersFromProject(projectId);
@@ -108,7 +117,13 @@ public class ProjectController {
     }
 
     @GetMapping("/edit/{projectId}")
-    public String showEditProject(@PathVariable int projectId, Model model) {
+    public String showEditProject(@PathVariable int projectId, Model model, HttpSession session) {
+
+        User currentLoggedInUser = (User) session.getAttribute("user");
+        if (!userService.isUserLoggedIn(currentLoggedInUser)) {
+            return "redirect:/profile/login";
+        }
+
         model.addAttribute("project", projectService.findProjectById(projectId));
         model.addAttribute("allUsers", userService.findAllUsers());
         model.addAttribute("assignedUsersIds", userService.findUserIdsAssignedToProjectByProjectId(projectId));
@@ -120,7 +135,13 @@ public class ProjectController {
     public String editProject(@PathVariable int projectId,
                               @RequestParam String projectName,
                               @RequestParam String projectDescription,
-                              @RequestParam LocalDate projectStartDate, @RequestParam(required = false) List<Integer> listOfUserIdsFromAssignedUsers) {
+                              @RequestParam LocalDate projectStartDate, @RequestParam(required = false) List<Integer> listOfUserIdsFromAssignedUsers, HttpSession session) {
+
+        User currentLoggedInUser = (User) session.getAttribute("user");
+        if (!userService.isUserLoggedIn(currentLoggedInUser)) {
+            return "redirect:/profile/login";
+        }
+
         projectService.updateProject(projectId, projectName, projectDescription, projectStartDate);
 
         projectService.removeAllUsersFromProject(projectId);
@@ -131,9 +152,9 @@ public class ProjectController {
 
     @GetMapping("allProjects")
     public String showAllProjects(HttpSession session, Model model) {
-        User currentUser = (User) session.getAttribute("user");
 
-        if (currentUser == null) {
+        User currentLoggedInUser = (User) session.getAttribute("user");
+        if (!userService.isUserLoggedIn(currentLoggedInUser)) {
             return "redirect:/profile/login";
         }
